refactor: implement GetTodaysDailyQuestProvider for daily quest retrieval and simplify DailyQuestService
