+++ backend/src/quests/quests.module.ts
@@ -4,6 +4,7 @@ import { DailyQuest } from './entities/daily-quest.entity';
 import { DailyQuestPuzzle } from './entities/daily-quest-puzzle.entity';
 import { DailyQuestController } from './controllers/daily-quest.controller';
 import { DailyQuestService } from './providers/daily-quest.service';
+import { GetTodaysDailyQuestProvider } from './providers/getTodaysDailyQuest.provider';
 import { PuzzlesModule } from '../puzzles/puzzles.module';
 import { ProgressModule } from '../progress/progress.module';
 import { UsersModule } from '../users/users.module';
@@ -16,7 +17,7 @@ import { UsersModule } from '../users/users.module';
     UsersModule,
   ],
   controllers: [DailyQuestController],
-  providers: [DailyQuestService],
+  providers: [DailyQuestService, GetTodaysDailyQuestProvider],
   exports: [TypeOrmModule, DailyQuestService],
 })
 export class QuestsModule {}
